Add file caching
The files were already cached (saved locally), but it was not tried to use them. This names files by their hash to ensure that no old file version is used. The side effect of this is that equal files are only downloaded and saved once.

This closes https://github.com/syncthing/syncthing-lite/issues/43

What's missing:

- delete local files which are no longer provided by the server
- delete local files which were not used for a long time
- delete all local files when the menu item to clear the cache is used